当无需取消的时候更新其DisplayListener

+++ src/me/xiaopan/android/imageloader/task/download/DownloadCallable.java
@@ -16,13 +16,7 @@
 
 package me.xiaopan.android.imageloader.task.download;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.*;
 import java.net.SocketTimeoutException;
 import java.util.Map;
 import java.util.WeakHashMap;
@@ -34,28 +28,15 @@ import me.xiaopan.android.imageloader.util.ImageLoaderUtils;
 
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
-import org.apache.http.HttpVersion;
+import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.conn.ConnectTimeoutException;
 import org.apache.http.conn.ConnectionPoolTimeoutException;
-import org.apache.http.conn.params.ConnManagerParams;
-import org.apache.http.conn.params.ConnPerRouteBean;
-import org.apache.http.conn.scheme.PlainSocketFactory;
-import org.apache.http.conn.scheme.Scheme;
-import org.apache.http.conn.scheme.SchemeRegistry;
-import org.apache.http.conn.ssl.SSLSocketFactory;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.apache.http.impl.conn.tsccm.ThreadSafeClientConnManager;
-import org.apache.http.params.BasicHttpParams;
-import org.apache.http.params.HttpConnectionParams;
-import org.apache.http.params.HttpProtocolParams;
 
 import android.util.Log;
+import org.apache.http.entity.BufferedHttpEntity;
 
 public class DownloadCallable implements Callable<Object>{
-    private static final int DEFAULT_CONNECTION_TIME_OUT = 2000;
-    private static final int DEFAULT_MAX_CONNECTIONS = 10;
-    private static final int DEFAULT_SOCKET_BUFFER_SIZE = 8192;
 	private static final String NAME = DownloadCallable.class.getSimpleName();
 	private static final Map<String, ReentrantLock> urlLocks = new WeakHashMap<String, ReentrantLock>();
 	private DownloadRequest downloadRequest;
@@ -92,12 +73,12 @@ public class DownloadCallable implements Callable<Object>{
 		
 		Object result = null;
 		int numberOfLoaded = 0;	//已加载次数
-		DefaultHttpClient defaultHttpClient = createHttpClient();
-		while(true){
+		HttpClient defaultHttpClient = downloadRequest.getConfiguration().getHttpClientCreator().onCreatorHttpClient();
+        while(true){
 			numberOfLoaded++;//加载次数加1
 			HttpGet httpGet = null;
-			BufferedInputStream bufferedfInputStream = null;
-			BufferedOutputStream bufferedOutputStream = null;
+			InputStream inputStream = null;
+			OutputStream outputStream = null;
 			try {
 				//发送请求
 				httpGet = new HttpGet(downloadRequest.getUri());
@@ -105,71 +86,49 @@ public class DownloadCallable implements Callable<Object>{
 				long fileLength = parseContentLength(httpResponse);
 				
 				//读取数据
-				bufferedfInputStream = new BufferedInputStream(httpResponse.getEntity().getContent(), ImageLoaderUtils.BUFFER_SIZE);
-				if(downloadRequest.getCacheFile() != null && ImageLoaderUtils.createFile(downloadRequest.getCacheFile())){
-					downloadRequest.getConfiguration().getBitmapCacher().setCacheFileLength(downloadRequest.getCacheFile(), fileLength);
-					bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(downloadRequest.getCacheFile(), false));
-					copy(bufferedfInputStream, bufferedOutputStream, fileLength);
-                    if(downloadRequest.getCacheFile().length() == fileLength){
-					    result = downloadRequest.getCacheFile();
-                        if(downloadRequest.getConfiguration().isDebugMode()){
-                            Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载成功 - FILE").append("；").append(downloadRequest.getName()).toString());
-                        }
-                    }else{
-                        downloadRequest.getCacheFile().delete();
-                        result = null;
-                        if(downloadRequest.getConfiguration().isDebugMode()){
-                            Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载失败 - FILE - 文件长度不匹配").append("；").append(downloadRequest.getName()).toString());
-                        }
+                inputStream = new BufferedHttpEntity(httpResponse.getEntity()).getContent();
+                if(downloadRequest.getCacheFile() != null && ImageLoaderUtils.createFile(downloadRequest.getCacheFile()) && downloadRequest.getConfiguration().getDiskCache().applyForSpace(fileLength)){
+                    // 如果可以缓存到本地
+                    outputStream = new BufferedOutputStream(new FileOutputStream(downloadRequest.getCacheFile(), false));
+                    copy(inputStream, outputStream, fileLength);
+                    result = downloadRequest.getCacheFile();
+                    if(downloadRequest.getConfiguration().isDebugMode()){
+                        Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载成功 - FILE").append("；").append(downloadRequest.getName()).toString());
                     }
 				}else{
-					ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-					bufferedOutputStream = new BufferedOutputStream(byteArrayOutputStream);
-					copy(bufferedfInputStream, bufferedOutputStream, fileLength);
-					byte[] data = byteArrayOutputStream.toByteArray();
-                    if(data.length == fileLength){
-                        result = data;
-                        if(downloadRequest.getConfiguration().isDebugMode()){
-                            Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载成功 - BYTE_ARRAY").append("；").append(downloadRequest.getName()).toString());
-                        }
-                    }else{
-                        result = null;
-                        if(downloadRequest.getConfiguration().isDebugMode()){
-                            Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载失败 - BYTE_ARRAY - 数据长度不匹配").append("；").append(downloadRequest.getName()).toString());
-                        }
+                    // 如果需要直接读到内存
+                    ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+                    outputStream = new BufferedOutputStream(byteArrayOutputStream);
+                    copy(inputStream, outputStream, fileLength);
+                    result = byteArrayOutputStream.toByteArray();
+                    if(downloadRequest.getConfiguration().isDebugMode()){
+                        Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载成功 - BYTE_ARRAY").append("；").append(downloadRequest.getName()).toString());
                     }
-				}
-				break;
+                }
+                break;
 			} catch (Throwable e2) {
-				e2.printStackTrace();
-				if(httpGet != null){
-					httpGet.abort();
-				}
-				
-				//删除文件
-				if(downloadRequest.getCacheFile() != null && downloadRequest.getCacheFile().exists()){
-					downloadRequest.getCacheFile().delete();
-				}
-				
+                if(httpGet != null) httpGet.abort();
+				if(downloadRequest.getCacheFile() != null && downloadRequest.getCacheFile().exists()) downloadRequest.getCacheFile().delete();
+
 				boolean isRetry = false;	//如果尚未达到最大重试次数，那么就再尝试一次
 				if(e2 instanceof ConnectTimeoutException || e2 instanceof SocketTimeoutException  || e2 instanceof  ConnectionPoolTimeoutException){
 					if(downloadRequest.getDownloadOptions() != null && downloadRequest.getDownloadOptions().getMaxRetryCount() > 0){
 						isRetry = numberOfLoaded < downloadRequest.getDownloadOptions().getMaxRetryCount();
 					}
-				}
+				}else{
+				    e2.printStackTrace();
+                }
 				
-				if(downloadRequest.getConfiguration().isDebugMode()){
-					Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载异常").append("；").append(downloadRequest.getName()).append("；").append("异常信息").append("=").append(e2.toString()).append("；").append(isRetry?"重新下载":"不再下载").toString());
-				}
-				
+				if(downloadRequest.getConfiguration().isDebugMode()) Log.d(ImageLoader.LOG_TAG, new StringBuffer(NAME).append("：").append("下载异常").append("；").append(downloadRequest.getName()).append("；").append("异常信息").append("=").append(e2.toString()).append("；").append(isRetry?"重新下载":"不再下载").toString());
+
 				if(!isRetry){
 					break;
 				}
-			}finally{
-				ImageLoaderUtils.close(bufferedfInputStream);
-				ImageLoaderUtils.close(bufferedOutputStream);
-			}
-		}
+			}finally {
+                ImageLoaderUtils.close(inputStream);
+                ImageLoaderUtils.close(outputStream);
+            }
+        }
 		return result;
 	}
 
@@ -187,26 +146,6 @@ public class DownloadCallable implements Callable<Object>{
         outputStream.flush();
         return completedLength;
     }
-
-    /**
-     * 创建一个HTTP客户端
-     * @return HTTP客户端
-     */
-	private static DefaultHttpClient createHttpClient(){
-		BasicHttpParams httpParams = new BasicHttpParams();
-        ConnManagerParams.setTimeout(httpParams, DEFAULT_CONNECTION_TIME_OUT);
-        HttpConnectionParams.setSoTimeout(httpParams, DEFAULT_CONNECTION_TIME_OUT);
-        HttpConnectionParams.setConnectionTimeout(httpParams, DEFAULT_CONNECTION_TIME_OUT);
-        ConnManagerParams.setMaxConnectionsPerRoute(httpParams, new ConnPerRouteBean(DEFAULT_MAX_CONNECTIONS));
-        ConnManagerParams.setMaxTotalConnections(httpParams, DEFAULT_MAX_CONNECTIONS);
-        HttpConnectionParams.setSocketBufferSize(httpParams, DEFAULT_SOCKET_BUFFER_SIZE);
-        HttpConnectionParams.setTcpNoDelay(httpParams, true);
-        HttpProtocolParams.setVersion(httpParams, HttpVersion.HTTP_1_1);
-		SchemeRegistry schemeRegistry = new SchemeRegistry();
-		schemeRegistry.register(new Scheme("http", PlainSocketFactory.getSocketFactory(), 80));
-		schemeRegistry.register(new Scheme("https", SSLSocketFactory.getSocketFactory(), 443));
-		return new DefaultHttpClient(new ThreadSafeClientConnManager(httpParams, schemeRegistry), httpParams); 
-	}
 
     /**
      * 解析内容长度
